+++ apps/i18n/common/en_us.json
@@ -2791,6 +2791,7 @@
   "sharingBlockedByTeacher": "Sorry, you do not have permissions to share this project. If you want to be able to share your project, please ask your teacher to enable sharing of App Lab / Game Lab / Web Lab projects for your section from the 'Manage students' tab in their dashboard. They can do this by adding the project sharing column from the Actions settings menu.",
   "sharingDisabled": "Sorry, this project is not available for sharing. If this is your project or the project of one of your students, please [sign in]({sign_in_url}) to your account to view the project.",
   "show": "Show",
+  "showAiInsights": "Show AI Insights",
   "showAllLessons": "Show All Lessons",
   "showAnswer": "Show answer",
   "showAnswers": "Show answers",
