+++ apps/src/templates/levelSummary/FreeResponseResponses.jsx
@@ -10,23 +10,11 @@ import React, {useEffect} from 'react';
 import {EVENTS, PLATFORMS} from '@cdo/apps/metrics/AnalyticsConstants';
 import analyticsReporter from '@cdo/apps/metrics/AnalyticsReporter';
 import {getFullName} from '@cdo/apps/templates/manageStudents/utils.ts';
-import experiments from '@cdo/apps/util/experiments';
 import i18n from '@cdo/locale';
-
-import FreeResponseAIEvaluation from './FreeResponseAIEvaluation';
 
 import styles from './summary.module.scss';
 
-const FreeResponseResponses = ({
-  responses,
-  showStudentNames,
-  eventData,
-  unitName,
-}) => {
-  const levelData = {
-    levelId: eventData.levelId,
-    unitId: eventData.unitId,
-  };
+const FreeResponseResponses = ({responses, showStudentNames, eventData}) => {
   const constructStudentName = response =>
     getFullName(response.student_display_name, response.student_family_name);
 
@@ -122,16 +110,6 @@ const FreeResponseResponses = ({
     </div>
   );
 
-  const AiEvaluationMVPUnits = ['csp4-2024', 'csp6-2024'];
-  const showAIAnalysis =
-    experiments.isEnabled(experiments.FREE_RESPONSE_AI_ANALYSIS) &&
-    AiEvaluationMVPUnits.includes(unitName);
-  const responsesForAi = responses.map(response => ({
-    studentId: response.user_id,
-    studentDisplayName: response.student_display_name,
-    studentWork: response.text,
-  }));
-
   return (
     <div className={styles.studentResponsesContent}>
       {pinnedResponses.length > 0 && (
@@ -225,12 +203,6 @@ const FreeResponseResponses = ({
             numHiddenResponses: hiddenResponses.length,
           })}
           type="gray"
-        />
-      )}
-      {showAIAnalysis && (
-        <FreeResponseAIEvaluation
-          responses={responsesForAi}
-          levelData={levelData}
         />
       )}
     </div>
